+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.4.0
+
+- added \left and \right
+- initial support for display style vs inline math mode
+- added glyph for \pm
+- added website that compares Dart-TeX with MathJax
+- sub and sup for matrixes are now vertically aligned correctly
+- improved glyph dimensions
+- fixing of minor bugs
+
 ## 0.3.1
 
 - the height attribute is now exported in the SVG data